+++ sdk/ai/Azure.AI.Projects/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 2.0.0 (Unreleased)
 
+### Other Changes
+* Added `finetuning` samples for operations create, retrieve, list, list events, list checkpoints, cancel, pause and resume. Also, these samples includes various finetuning techniques like Supervised (SFT), Reinforcement (RFT) and Direct performance optimization (DPO).
+
 ## 1.1.0 (2025-11-03)
 
 ### Other Changes